reducers and actions split into several files for convenience
