+++ src/components/app/app.jsx
@@ -2,7 +2,9 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
-import { getIngredients, addIngredient, closeModal } from '../../services/actions';
+import { getIngredients } from '../../services/actions/ingredientsActions';
+import { addIngredient } from '../../services/actions/constructorActions';
+import { closeModal } from '../../services/actions/modalActions';
 import AppHeader from '../app-header/app-header';
 import BurgerIngredients from '../burger-ingredients/burger-ingredients';
 import BurgerConstructor from '../burger-constructor/burger-constructor';
